src/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "semantic-kernel>=1.32.2",
    "azure-ai-agents>=1.0.1",
    "azure-identity>=1.23.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "aiohttp>=3.12.13",
    "azure-ai-inference>=1.0.0b9",
]
